workspace/TenderCutsApp/requirements.txt
# Appium should be installed
selenium
Appium-Python-Client
pytest
pytest-bdd
mysqlclient
requests